requirements.txt
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.3.0
scipy>=1.7.0
mysql-connector-python>=8.0.0
PyYAML>=5.4.0
SQLAlchemy>=1.4.0 